requirements.txt
# ═══════════════════════════════════════════════════════════════
# Postulator — Dépendances Python complètes (v1.0.0)
# Installation : pip install -r requirements.txt
# ═══════════════════════════════════════════════════════════════
# Note : greenlet doit être installé EN PREMIER sur Python 3.13
#   pip install "greenlet>=3.1.0" && pip install -r requirements.txt
# ═══════════════════════════════════════════════════════════════

# ── Web framework ─────────────────────────────────────────────
fastapi==0.115.0
uvicorn==0.30.6

# ── ORM & base de données ────────────────────────────────────
greenlet>=3.1.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
alembic==1.13.3

# ── Validation & settings ────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2

# ── Scraping ─────────────────────────────────────────────────
python-jobspy>=1.1.80
beautifulsoup4==4.12.3
httpx==0.27.2
fake-useragent==1.5.1

# ── IA locale (Ollama) ───────────────────────────────────────
ollama==0.3.3

# ── Parsing PDF / CV ─────────────────────────────────────────
pymupdf>=1.24.11
python-multipart==0.0.12

# ── Tâches asynchrones ───────────────────────────────────────
celery>=5.4.1
redis==5.1.1

# ── Utilitaires ──────────────────────────────────────────────
python-dotenv==1.0.1
tenacity==9.0.0
loguru==0.7.2

# ── Optionnel — export DOCX ──────────────────────────────────
# pandoc doit être installé séparément : sudo apt install pandoc
